backend/requirements.txt
# Meow Chat - OCR & Chat Application Requirements
# Python 3.10+ compatible

# 🌐 웹 프레임워크 (conda에 없음)
uvicorn[standard]==0.31.1
python-multipart==0.0.9

# 🤖 AI/ML 특화 패키지 (PaddleOCR 3.2+ 스택)
paddlepaddle>=2.6.0
paddleocr>=3.2.0
opencv-contrib-python>=4.8.0    # PaddleX 요구사항에 맞춰 유연하게 변경

# 📊 데이터 검증 & 처리 (conda pyyaml 보완)
pydantic>=2.5.0,<3.0.0
pydantic-core>=2.14.0
ruamel.yaml>=0.18.5              # conda pyyaml보다 고급 기능

# 🔧 OCR 특화 텍스트 처리 (conda에 없음)
pyparsing>=3.1.1
pyclipper>=1.3.0
python-bidi>=0.4.2              # 양방향 텍스트
arabic-reshaper>=3.0.0          # 아랍어 텍스트 처리

# 🎯 이미지 증강 & 전처리 (conda scikit-image 보완)
imgaug>=0.4.0
albumentations>=1.3.1

# 🔗 외부 서비스 연동 (conda에 없음)
modelscope>=1.28.0
huggingface-hub>=0.19.4
transformers>=4.36.0
aistudio-sdk>=0.3.0
onnxruntime>=1.16.3

# 🤖 LLM 클라이언트 (OpenAI)
# 고정 버전으로 의존성 백트래킹 지연 방지
openai==2.6.1

# 🤖 OpenAI Agents SDK
openai-agents==0.4.2

# 🤖 LangChain MCP 통합 (백트래킹 방지를 위해 고정)
langchain==1.0.3
langchain-community==0.4.1
langchain-openai==1.0.1
langchain-mcp-adapters==0.1.11
langgraph==1.0.2

# 🧩 Model Context Protocol 서버/클라이언트
mcp==1.19.0

# 🖥️ Streamlit 앱 런타임
streamlit==1.51.0
python-dotenv>=1.0.1
tiktoken==0.12.0              # 정확한 토큰 카운팅

# 🔎 선택적 벡터 스토어(설치 시 성능/품질 향상)
chromadb==1.3.0              # 설치 시 Chroma 백엔드 사용

# 🌐 웹 프레임워크 호환성 개선 (pip 백트래킹 방지용 버전 고정)
fastapi==0.120.2         # anyio 4.x 호환
starlette==0.49.1

# 🔧 의존성 충돌 해결을 위한 버전 고정
aiofiles==23.2.1         # gradio 4.19.0 호환
anyio>=4.5               # mcp, sse-starlette 호환
numpy>=1.26.4,<2.0       # gradio/paddlepaddle 양쪽 호환 (opencv는 경고만)
pillow==10.4.0           # gradio 4.19.0 호환
websockets==11.0.3       # gradio-client 0.10.0 호환
pypdfium2==4.30.0        # img2table 1.4.2 호환
zope-interface>=6.0      # datetime 패키지 호환

# 💾 고성능 데이터 형식 (conda 기본 보완)
ujson>=5.8.0                    # 고속 JSON
shapely>=2.0.2                  # 기하학적 객체
h5py>=3.10.0                    # HDF5 Python 인터페이스

# � HTTP 클라이언트 스택 (백트래킹 방지용 고정)
httpx==0.28.1
httpcore==1.0.9

# �🔒 암호화 & 보안 (conda에 없음)
pycryptodome>=3.19.0
bce-python-sdk>=0.9.0

# ⚡ 성능 유틸리티 (conda 기본 보완)
tqdm>=4.66.1                    # 진행 표시줄
filelock>=3.13.1               # 파일 잠금
fsspec>=2023.12.2              # 파일시스템 추상화
psutil>=5.9.6                  # 시스템 모니터링
rapidfuzz>=3.5.2               # 빠른 문자열 매칭

# 📦 타입 & 호환성
typing-extensions>=4.8.0
annotated-types>=0.6.0

# 🎯 딥러닝 지원 (conda pytorch 보완)
protobuf>=4.25.1,<5.0.0
six>=1.16.0

# 🔄 비동기 처리 (conda aiohttp 보완)
asyncio-throttle>=1.0.2

# 📏 이미지 품질 평가 (수정된 버전)
ssim>=0.3.0                     # 구조적 유사성 (최신 가능 버전)
pytorch-msssim>=1.0.0          # PyTorch SSIM 구현체 (더 나은 대안)